archive/projects_nested/screenwatcher_project/requirements.txt
# Framework
Django>=5.2,<5.3
djangorestframework>=3.14
django-filter>=23.0
channels>=4.0
daphne>=4.0
channels-redis>=4.0

# Database drivers
# mysqlclient>=2.1  # Removed due to missing system headers
PyMySQL>=1.0
django-mysql>=4.0

# Asynchronous tasks
celery>=5.2
django-celery-results>=2.5
redis>=4.6

# HTTP client
httpx>=0.25
aiohttp>=3.8

# MQTT support
paho-mqtt>=1.6

# Image handling
pillow>=10.0

# Rules Engine
json-logic>=0.7.0a0
casbin>=1.0
casbin-django-orm-adapter>=1.0

# Observability
opentelemetry-sdk>=1.0
opentelemetry-instrumentation-django>=0.39b0
drf-spectacular>=0.26
whitenoise>=6.5

# Utilities
pydantic>=2.0
python-json-logger>=2.0
django-environ>=0.10
pytest-django>=4.5
pytest>=7.0
pytest-cov>=4.0
freezegun>=1.2
flake8>=6.0
black>=23.0
locust>=2.15.0
django-cors-headers
django-unfold
mcp
uvicorn
elasticsearch>=8.0.0
